chore: 특수 문자 제거하는 로직 추가
