+++ features/record/index.ts
@@ -6,3 +6,4 @@ export * from './server-actions';
 export * from './store';
 export * from './styles';
 export * from './types';
+export * from './utils';
