+++ features/record/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './highlight-text';
+export * from './remove-special-symbols';
