+++ features/record/utils/remove-special-symbols.ts
@@ -1,0 +1,3 @@
+export const removeSpecialSymbols = (text: string) => {
+  return text.replace(/[^a-zA-Z0-9]/g, '');
+};
