+++ hooks/revalidateCurrentMemberInfo.ts
@@ -1,0 +1,7 @@
+'use server';
+
+import { revalidateTag } from 'next/cache';
+
+export const revalidateCurrentMemberInfo = () => {
+  revalidateTag('currentMember');
+};
